and rename the file as well
