+++ typescript/packages/common-charm/src/index.ts
@@ -1,4 +1,4 @@
-export { SpaceManager, type Charm } from "./charm.js";
+export { SpaceManager, type Charm } from "./space-manager.js";
 export { saveRecipe } from "./syncRecipe.js";
 export { buildRecipe, tsToExports } from "./localBuild.js";
 export {
