+++ typescript/packages/common-charm/src/space-manager.ts
